Refactor : 테스트 코드 위치 변경
검증 파트 article 과 user 분리
