+++ src/test/java/kr/codesqaud/cafe/validation/article/ArticleWriteTest.java
@@ -1,8 +1,7 @@
-package kr.codesqaud.cafe.validation;
+package kr.codesqaud.cafe.validation.article;
 
-import kr.codesqaud.cafe.dto.article.ArticleUpdateDTO;
+import kr.codesqaud.cafe.dto.article.ArticleFormDTO;
 import kr.codesqaud.cafe.validation.article.ArticleNewFormValidator;
-import kr.codesqaud.cafe.validation.article.ArticleUpdateValidator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,25 +10,24 @@ import org.springframework.validation.BeanPropertyBindingResult;
 import org.springframework.validation.Errors;
 import org.springframework.validation.FieldError;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.*;
 
 @SpringBootTest
-class ArticleUpdateTest {
+class ArticleWriteTest {
 
-
-    private final ArticleUpdateValidator articleValidator;
+    private final ArticleNewFormValidator articleValidator;
     private final String TITLE_ERROR_CODE = "required.article.title";
     private final String CONTENT_ERROR_CODE = "required.article.contents";
 
     @Autowired
-    public ArticleUpdateTest(ArticleUpdateValidator articleValidator) {
+    public ArticleWriteTest(ArticleNewFormValidator articleValidator) {
         this.articleValidator = articleValidator;
     }
 
     @Test
-    @DisplayName("업데이트 시, 제목이 비어있으면 에러 발생")
+    @DisplayName("새로운 글 작성 시, 제목이 비어있으면 에러 발생")
     void checkNoTitle() {
-        ArticleUpdateDTO article = new ArticleUpdateDTO("", "content");
+        ArticleFormDTO article = new ArticleFormDTO("", "", "content");
         Errors errors = new BeanPropertyBindingResult(article, "article");
         articleValidator.validate(article, errors);
 
@@ -39,9 +37,9 @@ class ArticleUpdateTest {
     }
 
     @Test
-    @DisplayName("업데이트 시, 내용이 비어있으면 에러 발생")
+    @DisplayName("새로운 글 작성 시, 내용이 비어있으면 에러 발생")
     void checkNoContents() {
-        ArticleUpdateDTO article = new ArticleUpdateDTO("title", "");
+        ArticleFormDTO article = new ArticleFormDTO("", "title", "");
         Errors errors = new BeanPropertyBindingResult(article, "article");
         articleValidator.validate(article, errors);
 
@@ -51,11 +49,12 @@ class ArticleUpdateTest {
     }
 
     @Test
-    @DisplayName("형식에 맞는 업데이트 작성 시 에러 없이 정상 등록")
+    @DisplayName("형식에 맞는 글 작성 시 에러 없이 정상 등록")
     void writeArticle() {
-        ArticleUpdateDTO article = new ArticleUpdateDTO("title", "contents");
+        ArticleFormDTO article = new ArticleFormDTO("first", "title", "contents");
         Errors errors = new BeanPropertyBindingResult(article, "article");
         articleValidator.validate(article, errors);
         assertThat(errors.hasErrors()).isFalse();
     }
+
 }
